move preprocessing to mha2nnunet
